Fix a bug in the specialist guidance pagination js
The main fix here is to prepend the hash to the regexp we use to find
the link that we want to make active.  Without the hash, it was possible
to get multiple matching links, e.g. links to two pages, "page" and
"another-page" would both be found when a user clicked on "page".

In order to write a failing test for the problem I've had to change the
structure of the javascript slightly, so that I could control when the
pagination occurred.  To do so, I've created the GOVUK.paginateDocument
function that is called from the on_ready.js script.  I followed the
mainstream convention of using the GOVUK namespace and of initializing
the GOVUK object.

I've had to disable Sinon's Fake Timers so that I can use setTimeout in
my javascript tests.  If Sinon's Fake Timers are enabled (which they are
by default) then setTimeout never actually returns.

Finally, I've had to set the viewport size in PhantomJS so that the
pagination actually occurs (it's only applied when the window size is
greater than 768 pixels wide).  I'd like there to be a better way of
toggling the pagination behaviour so that we don't have to rely on the
window size but this'll do for now.