Rename collections to series, to better match usage.
After a few conversations, we've decided to produce with the concept
of a 'series', which is a set of publications related strongly
through time. For the moment, a publication can only be a member of a single series; this simplifies interface decisions and there's no
strong driver to put a document in multiple series at the moment.