Alternative format provider required if attachments present
This had a knock-on effect on many tests which created publications with
attachments but didn't set alternative format provider.

I changed the factories to add traits called :with_attachment and
:with_alternative_format_provider to support these scenarios.

I also bumped factory girl to the latest 2.x series as I had hit a bug
in 2.2.0 with using traits containing associations.