Fix attachments on Consultation responses
They were broken due to the addition of validation rules requiring
alternative format provider to be set when attachments are present.
These validation rules didn't work on consultation responses because
they were delegating the alternative_format_contact_email to their
parent consultation.

We couldn't find a reasonable way to continue supporting this behaviour,
so we've made consultation responses responsible for knowing their
alternative format provider.

We've also added tests relating to attachments on the admin
constultation responses controller. In adding these tests, we noticed
that there was a bug:- that on a validation error the new attachment
form is not displayed.