Prevent flickering on a page with short content
When scrolling, this feature would sometimes cause flickering. When the
sidebar floated (and was removed from document flow), it stopped taking
up space. This caused the overall page height to be recalculated often
scrolling the upper part of the page down, and thus moving the top of
the floating sidebar above the point where it should be floated, causing
the sidebar to be un-floated again.

I worked around this by inserting a fixed size invisible 'shim' div in
the place where the floated sidebar was removed.