Don't raise 500 errors if author or organisation filters aren't valid
The behaviour now is to redirect to either the remembered or default
filter options, just as if no options had been specified.  This
might not be the absolutely most user friendly thing to do, but as
this error is only likely to occur if urls are hacked by admins, I
believe it is sufficient.