Fix a bug with URI parsing in Guidance munger
CGI::Parse on an entire URL only behaves as you'd expect if the
key=value pair isn't the first in the query string. Instead we need to
use the URI library to get the query string and then parse it.

This also adds better reporting via Sets to provide uniqueness.