Move content around for new consultation designs
Moved consultation response attachments inline into consultation page so
they can easily be downloaded.

Moved the content of _document_content into show as it was only used in
one place. If the show is refactored to be shared with other views later
this can be re-split.

Moved consultation participation out of the full width attachment
partial as it will only be used on consultation#show. It is eaiser to
find in the show template.

Fixed _attachment_full_width to use local document rather than the
instance document.