Make the Images test more generic
This test should be focussed on testing the behaviour in the
Edition::Images module.  Following on from the approach we use in
alternative_format_provider_test.rb, I've introduced a new Edition
subclass that includes nothing more than the Images module.  Prior to
this, we were actually testing Policies with no guarantee that we
weren't testing Policy specific behaviour.

I wonder whether it'd be possible/easy to add test class specific
factory-girl factories to these types of tests, so that we can remove
some of the duplication introduced by our having to manually build the
objects required. That can wait for now, though.