Allow unpublished editions to be previewed
This feature is only available to users who are signed in via
GDS::SSO. If they pass a "preview=ID" param to a document and are signed
in, an arbitrary Edition of the viewed document will be fetched by the
ID passed.

The URL helper for this additionally adds a `cachebust` param to the URL
to ensure that any frontend caches will be correctly bypassed.