Remove images from Consultation Responses
I've introduced a NullImages module that implements some of the public
interface of the Images module, notably the 'reader' methods.  This
allowed me to reduce the number of changes I had to make to the
application as we can continue to assume that all editions respond to
`#images` and `#lead_image`.  I think I quite like this approach even
though I think it's slightly different from what we've done in the past.