Add the summary field to Consultation responses
We're in the process of changing the way Consultation responses are
displayed.  They'll eventually only have an optional summary rather than
an optional summary and required body, but this is just a step on that
fine journey.

*NOTE* You can now add a summary to a Consultation response that'll show
in the admin interface but *not* in the frontend.  Irrespective, this
behaviour is not tested and so can only be considered incidental.  I'll
be displaying the summary in the frontend soon enough.