Remove the body from Consultation responses
You could previously add a summary to a Consultation response but this
commit promotes that to be the primary block of text we show in the
front end.

I updated the consultation_responses factory to explicitly set the body
to nil.  This was to force some test failures that we'd see in the app
now that we can create responses without bodies.  Before setting the
body to nil it was being set by the its parent edition factory to
'edition-body'.

I've added a migration to move the data from the body to summary for
existing consultation responses.  This wasn't strictly necessary as
we're going to be replacing all this data but it still felt like the
right thing to do for now.