Refactor stick at top when scrolling
The reason for this was to fix a bug where if the content on the page
changed after page load the stick at top would break. Howerver, I saw
this as an opertunity to refactor this into the new style that we are
adopting for javascript functionality.

The reason for using `call(this)` is that it can then be called from a
non-`window` scop and not break anything.