Display the consultation response summary
We "removed" (in as much as you can in STI land) the body from
Consultation responses in a8d77b9bd17a154e9c83c1b4db871e6e1d9b3dfe, and
now use the summary where required.

*NOTE* In adding the tests for this behaviour, I discovered that
we only display the title of the consultation response when there aren't
any attachments.  If there are attachments, we display the title of the
attachment rather than the title of the response.  It's not obvious to
me whether this behaviour is by design or not.  I also find it odd that
we display the summary in an element with a class of 'extra-description'
in the _attachment partial.  I've probably added to that confusion by
using a class name of 'summary' in the show template.