Make warden's behaviour in cucumber test match development.
We were seeing that in cucumber, despite 'logging in' our editor
was not able to visit the preview of a document[1]. It seemed as if
warden was not restoring the user from session information without
a call to `warden.authenticate?`, which wasn't required when
running the server in development mode and the
`GDS_SSO_MOCK_INVALID` flag set on.

By including the Warden test helpers, it seems to work. Fingers
crossed that this isn't just weird voodoo.

[1]: Run `cucumber features/preview-unpublished-editions.feature -n "Unpublished editions can be previewed"` before this commit.