Automatically add change note when response published
Consultation responses can now only be seen when you view the parent
consultation (i.e. they don't exist in isolation any longer).  As such,
we want to add an entry to the change history to state when a response
has been published.

I don't feel entirely comfortable with this approach for two reasons.
The first is that I'm sure this code could be clearer if we were
modelling Consultation responses and possible the change history,
differently.  The second is that I've introduced an in-memory sort which
may or may not cause us performance problems in the future.

*NOTE* I renamed the variable `last` to `oldest_change` as I kept
getting confused by what `last` meant when I was reading the code.