Associate attachments with Consultation responses
We should now be in a position to build responses and their associated
attachments directly from the Consultation.  The next step will be to
migrate existing Consultation Responses to this format and add the
relevant fields to admin Consultation form.

It feels slightly odd to use a full model
(ConsultationResponseAttachment) when a habtm relationship would do the
trick, but I think habtm relationships are somewhat out of favour.