Allow users to log out.
I've used the admin_users_path in the integration test, because the
admin_root_path requires two redirects before it actually renders
a page, and this may change in the future.

I would've preferred to write a cucumber test, but this is
prohibitively hard because we can't follow the redirect to the
sign-on server.