Add series information to publication index
This displays the information of a series the document is in in the
index and via the Ajax updating mechanism.

The Javascript to present this information is getting a bit hairy and
led to the decision to send back the text as a string rather than
URL/name pairs in a JSON object, but since Neil wanted to try it out
quickly I did the simplest thing.