Revert "Add validation to Consultation Response Attachments"
This reverts commit de48aa01fffbbae22bd4ffbd48a8468a79ad961e.

The presence of this validation makes it difficult to build a nested set
of objects (from the consultation to attachment through response and
consultation_response_attachment) so I'm going to revert it for now.  It
can always be added once we can actually create these objects through
the web interface.