Add responses to admin consultation form
As of this commit, we have two ways of creating 'responses'.  The
original method of creating Consultation Responses that are themselves
Editions, and the new method of creating Responses.  I'll soon be
migrating the existing Consultation Responses to Responses and removing
the ability to create the traditional Consultation Responses.

I've added the basic fields for creating attachments with the response
but plan to replace this with the attachment_fields partial (or parts of
it) in a later commit.

I don't like the `if @edition.is_a?(Consultation)` condition in the
editions/show template but I'm going to live with it for now.