Re-use attachment_fields partial for response attachments
Re-using this partial gives us all of the additional, optional,
attachment fields for free.  I've amended the response controller test
to assert the presence of these fields because at the moment they're
only visible because `Consultation#allows_attachment_references?`
returns true.

I've had to change the attachment_fields partial to allow us to define
whether or not we should be displaying the inline attachment markdown
help snippet.  This is so that we can continue to display the help for
the attachments on the consultation but not for those attachments on the
response.