Use the HTML validation from Govspeak to defend aganist XSS.
When a model includes `validates_with SafeHtmlValidator`, every
field that is changed on that model will be passed through HTML
validation to ensure that no script tags or other errant HTML is
present.

I've also added tests for the key fields we are concerned with; I
suppose technically we could add tests for every class and every
field, but that felt like overkill.