Allow empty response attachments
This should've been part of the previous commit
(fb7cfc8d48e527ec3d5dff71203c6c1b5793a46c).  Before this fix, creating a
new consultation would complain that the attachment title and file were
missing, irrespective of whether you were trying to add a response
attachment or not.  This is because the 'accessible attribute is always
sent.  I've copied the logic from the Attachable mixin to deal with this
attribute.

There's a failing feature in the last commit:
editing-draft-consultations.