Remove the actions to create/view consultation responses
We're moving to a world where consultation responses are created as part
of the consultation itself.  Before this commit you could still create a
Response (the new type) and a Consultation Response (the old type).
While you can still create a Consultation Response if you know the URL
this'll prevent people navigating to that form through the interface.

The next step is to migrate all the Consultation Responses to Responses
and remove all the old logic completely.

*NOTE* I've removed two features in this commit because they were
describing what happened when you create Consultation Responses.
There's probably an argument to say that I should've replaced them with
something that describes the creation of a Response in the new world,
but I don't have the energy.