Introduce simple specialist guidance API
This allows clients to request /specialist/api/specialist/<id>.json
to get a JSON representation of specialist guidance.  The format of
the JSON is based on that provided by the govuk_content_api, and it
is the eventual intention that all public requests for these pages
come through the content API.

The initial approach was to build on top of the existing specialist
guidance controller, but this was rejected because this controller
was already generating JSON used by the main web site (for infinite
search and similar).  Introducing a new controller and routes makes
the distinction between the two clearer.