Deny WAP requests to public controllers with 406
Google is crawling us with a WAP-like browser, which causes fatal errors
due to missing response formats.

Additionally, as our site will not function well on WAP connections due
to the large size of assets etc. we serve, we should deny the
connection until a better site is in place.

Note that using a blanket:

    respond_to :html

in the controller doesn't work for some reason, hence the filter.