Include the _response_info status hash, as in mainstream.
I would've like to contain the `_response_info` generation within
the presenter, but it doesn't seem like there's a trivial way to
indicate the response code for the controller within a custom
presenter like this.

Note also that the index *always* returns "ok"; I don't think there
is a scenario where the index would return something like "not
found" at the moment.