The API should respond with HTML without a govspeak wrapper div.
I think the govspeak helpers are ripe for refactoring into a
class which can be structured and tested a bit more clearly; at the
moment it's a soup of methods calling other methods. I had wanted
to simply pass an option (`wrap_in_govspeak_div: false` maybe) to
the method calls, but I lost confidence in how fast this would be
given the multivarious different entry points to govspeak rendering.