Replace use of ConsultationResponse with Response
I haven't removed the Consultations Responses yet as I still need to
copy the data from those responses to the new Response objects. As soon
as the data has been removed I'll be able to remove all the other code
that's hanging around to support Consultation Responses.

I've kept as similar an API as possible for Consultation in an attempt
to limit the amount I had to change in this commit, which is why we
still have `#published_consultation_response`, `#response_published?`
and `#response_published_on`.

There's not currently any validation to ensure that you choose an
alternative format provider if you add a response with an attachment.

I've had to amend the json representation of a consultation.  The only
failing test was a functional test so I'm hoping that I haven't
inadvertantly broken something that relied on the previous json
representation.