Introduce common timestamp for sorting editions.
I want to be able to sort a mixture of publications *and* consultations
according to their primary timestamp. However, the primary timestamp for
publications is different than for any other edition types.

This change means that a copy of the relevant primary timestamp is saved
into the new `timestamp_for_sorting` column. This new column can then be
used for sorting the mixture of editions that I described above.

I did consider using the `COALESCE` MySQL function in the query instead
of introducing the extra column, but it appears that this would have
meant the query could no longer make use of indices for the timestamp
columns. Since performance on this page is already a concern, I decied
to go with the extra column as described above.