Include alt format provider in controller_attributes
If the class has an association with alternative_format_provider then
the controller_attributes will automatically include a reference for
alternative_format_provider.

This is a bit icky as this is hidden in the tests, it might be good to
refactor this out to make it explicit later, but this is a quick way to
make tests more flexible for now.