Apply Attachable tests to SupporingPagesController
The tests are made more flexible so that they determine the name of the
controller attributes and join tables from the model under test.

Applying these tests to the supporting pages controller exposed a number
of bugs around attachments in the supporting pages controller which are
fixed by this commit.

I also noticed a bug when displaying a lock conflict on editing a
supporting page. The bug was due to the fact that two forms were shown
side-by-side on the same page with the same id, and the 'preview'
javascript code got confused in this case. I've fixed this by renaming
the second 'disabled' form on this page so that it has a different ID.