Make publications index work in JSON format with consultations.
* By adding `should_return_json_suitable_for_the_document_filter
:consultation` to the `PublicationsControllerTest`, this forces us to
make the JSON version of the action work for consultations. Before this
change if any consultations were returned, the JSON version of the
action failed. This in turn meant that infinite scroll was broken when
any consultations were returned.

* Unfortunately I've had to combine `PublicationPresenter` and
`ConsultationPresenter` into a single `PublicationesquePresenter` which
contains evil `case` statements. This is used by
`PublicationsController::PublicationesqueDecorator` to decorate all of
the returned models. I couldn't find a nicer way of doing this that
preserved the ActiveRecord::Relation behaviour of the returned
collection of models. It might be possible to improve on this by some
radical re-structuring of `DocumentFilterJsonPresenter` and subclasses,
but I'm not sure.

* I've removed `Whitehall::DocumentFilter#each_document_with_index`
since we can now simply use the decorated collection that
`PublicationsController::PublicationesqueDecorator#documents` returns.

* I've also change `PublicationFilterJsonPresenter#document_hash`
to use the presenter methods which was really where the main problem
was, because this code didn't work with consultations.

* I've also had to add a slightly unfortunate tweak to
`PublicDocumentRoutesHelper#model_name` when it is passed a presenter
instead of a model. This is necessary, because
`DocumentFilterJsonPresenter#document_hash` calls
`PublicDocumentRoutesHelper#public_document_path` and it would be
awkward to change this in just the `PublicationFilterJsonPresenter`
subclass. This way it's just as safe to pass a presenter decorating a
model as it is to pass the model itself.