Give the attachment fieldset a unique ID
This makes the multiple_file_upload javascript Just Work when the
'attachment_fields' partial is included multiple times on the same page.
Previous to this, only the first instance of the partial would have the
multiple upload functionality applied, as the javascript relies on the
ID of the fieldset being unique.

I feel a bit dirty not adding any tests for this but I didn't want to
add a Cucumber test and a functional test (ensuring that the two
instances had different IDs) didn't feel great either.