A guide may exist in a published state.
It's possible that a guide with the same title already exists, and
is in a state where it cannot be saved i.e. the first guide with
that title is in the "archive" state.

To work around this, I have changed the importer to find the *last*
guide with the same title, which should be either published or
draft, and then if necessary create a new draft edition to be
updated with the new content.

Since it's also possible that the guide won't be saved properly, we
can also report this.