Add screenshots for organisations joining or exempt from GOV.UK
These can be generated again using the `organisation_screenshots`
task. Pass an `ORGANISATION=<slug>` parameter to generate a
specific organisation screenshot. I anticipate that this will be
run on a developer's machine, not on the servers.

I decided to implement this using files in the repository, rather
than storing images in S3 or the database, because it was simpler
than adding another attachment model, and these files should not
need to be managed by admin users.

I've also cleaned the data in preview and production based on
running the task locally.