Add Atom feed containing latest activity in a topic
I copied the tests from those in home_controller_test and copied the
implementation from home/show.atom.builder. One difference is that I've
tried to keep the tests away from the database as much as possible.  I
have had to create (i.e. persist) the document object in two of the
tests because using a stub resulted in an unrecognised routes error.

I've left out the top-level summary attribute of the feed as it
wasn't obvious to me what it's used for.

I'm displaying the Edition#summary for each entry, in addition to
the title and content fields displayed in the home/show.atom and
publications/index.atom templates.

I've moved the logic contained in
TopicsController#recently_changed_documents to the Topic model and unit
tested it separately.  This allowed me to stub the method in the
controller and both simplify and cut down on database access in the atom
feed tests.