Limit the Topic feed to 10 entries
We don't appear to have a standard for the number of entries to show in
an atom feed (the site atom feed is limited to 10 entries, publications
are limited to 20, and policy activity doesn't appear to be limited).
I've decided to go with 10 for now but we can obviously change this as
and when necessary.