Add format to API id urls.
While I think this is the wrong thing to do (ids should never
include the format), I'm doing it to be consistent with the rest of
the content api.

I think I will write an API browser or similar to try and avoid
having to do this and yet keep the API browsable.