DRY up atom feed link generation
I've introduced the link_to_feed helper to DRY up the atom feed link
generation markup. *NOTE* The empty span element with a class of
'feed-icon' didn't appear to be necessary so I've removed it.

I've changed the atom feed link on the homepage so that it's consistent
with the links on the publications index and policy activity stream, and
I've added a test to ensure that the link is present too.