Tweak layout of org edit page.
The list of featured documents is only relevant to the 'documents' tab,
we couldn't see a reason why it's pulled out and visible irrespective of
which tab is open.

The 'view on website' link feels like it needs to be associated with
the organisation name to give it context.

We renamed the 'documents' tab to 'featured documents' as that's the
primary activity you'll do on that tab.