Remove unused 'feature' functionality
The Edition#featured column isn't used anywhere so we've just removed
it.  We had to update a Consultation controller test that was setting
the 'featured' flag even though it didn't have any bearing on the
outcome of the test.

The Topic#featured column doesn't exist so we've removed some unused
code from that class.