Require change note on save rather than on publish
This gives a better user experience because the person who is actually
making the change is requested to enter the change note, rather than
expecting the second pair of eyes to enter a change note
retrospectively.

There are some edge cases around deleting a draft edition which didn't
have a change note, or deleting a published edition when it's the only
edition on a document. These are all covered by unit tests in
publishing_test.rb.

This change had side effects in the fixture set-up code in quite a few
tests (since change notes are now required at an early stage of the
documents lifecycle).