Add scheduled state and scheduled_publication date to edition
This includes the addition of a new state and state transition rules in
the state machine, and rules about publishing/scheduling in the
publishing mixin.

A submitted publication can be scheduled if it has scheduled_publication
date. A scheduled publication can be published once its
scheduled_publication date has passed.