Remove the option to paginate specialist guides.
As a result of the user feedback, it's been decided that having the
specialist guides as a single page will make it easier for users to
orient themselves and get a better idea of the length of the document.

There appears to be some non-pagination-related functionality in
`paginate_document.js`. I've left this functionality in place, but I
plan to move it into more appropriately named functions in a later
commit.