Extract Cucumber-specific gems into their own group.
And only load them when running Cucumber features.

I did this partly to break up the monolithic Gemfile and make it clearer
why some of the gems are actually needed, but also in the (perhaps vain)
hope that it might speed up some of the other test suites given they
will no longer load these gems.