Ensure all edition_organisations are valid.
In d70ac6266753464c7bd11fa5c0de6164f53d9cda, validation was added so
that a featured edition_organisation has to have an associated image.
Since none of the existing edition_organisations will have an associated
image, I've elected to mark all of them as unfeatured. This will mean
that they will all need to be marked as featured again, but this time an
image and alt_text will be required.

I did consider copying over data from NewsArticle#lead_image, but it's
been possible to feature any document type for a while (not just
new articles which are the only ones to have a `lead_image`), so I
thought it was simplest just to unfeature them all.