Remove the concept of a lead image.
News articles used to have a lead image, used when featuring them
in the carousel.  Now that all edition types can be featured, and
featuring includes adding an image, this is redundant.

Articles that have an existing lead image retain their images, but
if this image is to be shown, the markdown will need to be edited
to attach the image (just like any other document type).  All news
articles will be reviewed before going live to make changes like
this.