Allow filtering of publications by type.
Because PublicationType instances aren't really in the database, the
mechanism is slightly different but not problematically so.

As part of this I've increased the height of every filter because it
was the simplest way to proceed. However, it shouldn't be too hard
for the filter height to adapt according to how many controls are in
it.