Attempting to fix the failing build on CI.
It looks like the build is failing because of some interaction in
the cucumber tests with capybara-webkit. The exceptions being raised
are like:

    Capybara::Driver::Webkit::NodeNotAttachedError (Capybara::Driver::Webkit::NodeNotAttachedError)
    ./features/step_definitions/filtering_steps.rb:21:in `/^I should see that the (next|previous) page is (\d+) of (\d+)$/'
    features/filtering-policies.feature:35:in `Then I should see that the next page is 2 of 2'

After doing a bit of research, I found [this issue][1] and [this thread][2], which indicate that we shouldn't be using `wait_until`,
because the normal `find`-based DOM selectors will automatically
wait until their element appears in the page (or they time out).

Working under the assumption that avoiding using `wait_until` might
also avoid DOM-javascript interaction issue is causing the
exception, the rest of the changes are to support the removal of
the `wait_until` calls in the cucumber steps,

Because capybara-webkit runs in a separate thread from the server,
we quickly hit database problems where the connection is not ready at the point where another SQL request is sent ([`Mysql2::Error:
This connection is still waiting for a result, try again once you
have the result...`][3]) or validations fail.

The solution to this is to enable truncation (which doesn't seem to
significantly slow down the cucumber task in my local testing), but
also implement [a 'monkeypatch'][4] to ensure that all connections
to the database use a single, locked connection (via the
`connection_pool` gem).

It's quite possible that this won't help on the CI machine; I cannot
replicate the actual build failure locally so I cannot be certain.
If this doesn't fix the build, I think we should revert this change
and reconsider how these tests run.

[1]: https://github.com/jnicklas/capybara/issues/691
[2]: https://groups.google.com/forum/?fromgroups=#!topic/ruby-capybara/qQYWpQb9FzY
[3]: https://github.com/brianmario/mysql2/issues/99
[4]: https://github.com/brianmario/mysql2/issues/99#issuecomment-6766570