Ensure that scheduled_publication dates are in the future.
This revealed a bug in the scheduled publication form, which can be
replicated prior to this commit by submitting an invalid
publication; in the form showing the errors, the scheduled
publication check box will be checked. This is because the
condition was checking for the presence of *any* value in the
`scheduled_publication_active` parameter, rather than the "active"
value (`1`).