Add back to contents jump link
Now that there are no split pages there was a need to get back to the
contents quickly. A jump to contents link should now apear when the
navigation has scrolled out of view.

There is also now a jump to footer link in the navigation to get to the
related guidance.