Add rake tasks for publishing scheduled editions
Had to tweak #reason_to_prevent_publication_by to explicitly address the
rules around publishing scheduled editions.

Only require scheduled_publication to be in the future in the 'draft'
and 'submitted' state. I allow it to be in the past in 'rejected' state
because you may reject an edition and we don't want to nag people about
the scheduled_publication date in that case, it's up to the author or
departmental editor to set the correct scheduled_publication when
submitting it and scheduling it.