Ensure specialist guides always have a mainstream category.
* Add validation to ensure presence of
`SpecialistGuide#primary_mainstream_category`.

* Add migration to assign all existing specialist guides to an arbitrary
mainstream category. I've chosen the category with the first title
when arranged in alphabetical order. The idea is that the editors will
be able to choose a better primary mainstream category later, but this
keeps all the specialist guides valid so they can be edited and saved
successfully.

* Note that since every specialist guide used in a test now has a
primary mainstream category, we need to stub the content API in more
places including within Cucumber features.