Show "this replaces businesslink" for guides that do.
This will need some design attention, because at the moment this
content is "below the fold", and I think in order to meet the goal
of reassuring readers that they've arrived at the right place, it
will need to be more prominent.

I have duplicated the business link logo into this codebase, because
I was not entirely sure how to re-use images provided by the static
application (for mainstream, it uses a template that's actually *in*
the static codebase). Feel free to change this if there is a better
way.