Update Rummageable to fix broken search on preview
A recently deployed change to Rummager
(https://github.com/alphagov/rummager/pull/26) caused search to fail in
preview.  That change removed the /government prefix so we've had to
update our config to remove the prefix too.

In hindsight, I'm not entirely sure that I needed to update
Rummageable to get search working on preview again.  I'm happy to
leave the new version there as we'll need it when we switch to elastic
search in the near future.