Upgrade to latest version of gds-api-adapters
We need the latest version of the GdsApi::Rummager wrapper so that we
can connect to Rummager instances served using SSL.  This problem, of
trying to connect to Rummager using SSL, was inadvertently introduced in
fb98c4452b7cb21050c987bf271a5fecdde7b943 when the version of Plek was
updated.

This *should* fix the problem of search not working in preview.  The
error we're currently seeing in preview is:

    GdsApi::Rummager::SearchServiceError: 400
    400 The plain HTTP request was sent to HTTPS port