Consistent precison between UI and cron for scheduled publishing
Use the same minute increment both for the time selector dropdown and
the cron job schedule.

Assuming all our clocks are in sync, this should ensure a reasonable
level of accuracy in the actual time of publication. The cron command
should be invoked at the exact time, and will take a second or two
before it queries the database for editions due to be published.