Imported guides require a primary mainstream category.
Interestingly, the rake script completely explodes unless we
explicitly require the `Edition::MainstreamCategory` model, to
disambiguate it from the `MainstreamCategory` class.

I also found a bug in the reporting of the script, which might
explain why I previously thought a bunch of guides existed; we were
asking if a guide was a `new_record?` after it had been saved, when
it should never be 'new'.