Add breadcrumb metadata to search index for detailed guidance.
I would have preferred to use a single key to hold all this metadata,
but the strictures of Rummageable (client-side) and Rummager
(server-side) validation means that this would have been very awkward.
So for now it's more straightforward to re-use the "section" &
"subsection" keys used by mainstream and to introduce a new
"subsubsection" key.

This data is obtained by parsing the detailed guide's mainstream
category parent_tag and using its slug. The former in particular seems
rather brittle, but it seems like the least worst option at the moment.