Handle scenario where Content API has no metadata.
This was causing an exception when the metadata for the detailed guides
on preview disappeared for some reason. It was no longer possible to
view the page.

In the longer run, it might be better to fix this in the Content AP
client library, but this solves the problem for us now.