Introduce a quarantined file storage class.
The quarantined file storage class stores uploaded files in one
location, and reads them from a different location.  The intention
is that another process will be responsible for moving files between
the two locations, at which point they will be checked for viruses
and other nastiness.

This class is not yet plumbed in, as doing so will break things.