Remove bad code.
Using `zoom: 1;` isn't inherently offensive by itself. It is a good way
of enforcing `hasLayout` on elements that need it. However, when married
with the `*` selector it will be adding `hasLayout` to elements we don't
necessarily want layout. This can cause strange behaviour and make
things a pain to debug.

In this case we seem to only actually require it on one element.
Targeting that element would be a million times better and the right
thing to do.