WIP: Redesign news article pages
I have moved all of the content out of partials for the moment so that
we could easily move content around and mock things up without effecting
other pages. If this is the style we want to move forward with then we
will move the new markup into the partials.