Don't limit recent items in templates
This refers to 7a90484. Limiting should be done in the controllers
rather than the views.

While working on this, I discovered that publications could appear in
the same place more than once due to the design of the page, so I had to
add a prefix to the `recent_items` rendering to prevent the same ID
appearing twice in the page. I've added test helpers to achieve this and
make the tests better.