Explicit index when indexing ministers from an observer
Before this observer didn't specify an index, which meant it went to the
"primary" index, which was of course "mainstream". This is why Ministers
are showing up in the main search index.

I couldn't figure out a nicer way of handling this than hard-coding the
index in - ideally it'd use Searchable::Index on the instance being
passed in, but I can't figure out how to do that.