Migration to upload DFT news
This fixes the association with ministers.

I tried to make it stand-alone with respect to models, but the migration
relies on too many behaviours of the associations between edition and
policy - it seemed like too much work to replicate all of this in the
migration (in particular the updating of counters on the associations).

I'm going to investigate an alternative method of doing data migrations
so that we don't have to use schema migrations to make data changes.