Add a data migration mechanism
Using regular migrations to modify data is problematic. The main problem
is that data migrations generally do a lot of manipulation of models,
and ideally would depend upon model logic to enforce validation rules
etc. However, database schema migrations ideally should be de-coupled
from model logic to ensure that they can run in the future even when
model logic has changed.

The assumption of data migrations is that they'll generally only be run
on production databases and the choice of when to run them will be
controlled manually. There's no attempt to check the version of the
database schema present when running a data migration, it's expected
that that will be done manually.

A data migration follows a similar structure to a regular migration.
Create a file in db/data_migration/ with a name <datestamp>_<label>.rb

A data migration is instance_evaled, so you don't need to inherit any
classes in the file, just write ruby code and feel free to refer to any
model objects you need.